Split main routes into feature modules
