+++ app/main/views/__init__.py
@@ -1,0 +1,1 @@
+"""Feature-oriented route modules for the main blueprint."""
